+++ Aphelion/HBAPTutorialViewController.m
@@ -63,7 +63,7 @@
 }
 
 - (void)_setScrollViewContentSize {
-	unsigned numberOfPages = 3;
+	NSUInteger numberOfPages = 3;
 	
 	_scrollView.contentSize = CGSizeMake(self.view.frame.size.width * numberOfPages, self.view.frame.size.height);
 	_scrollView.contentOffset = CGPointMake(_scrollView.contentOffset.x, self.navigationController.navigationBar.frame.size.height);
